Fix dtype

+++ test_tmdb_etl.py
@@ -57,13 +57,13 @@ def test_find_by_external_id():
     df = pl.DataFrame({"imdb_id": ["tt1630029", "tt14269590", "nm3718007"]}).lazy()
 
     df2 = df.with_columns(tmdb_find(tmdb_type="movie", external_id_type="imdb_id"))
-    df3 = df.with_columns(pl.Series("tmdb_id", [76600, None, None]))
+    df3 = df.with_columns(pl.Series("tmdb_id", [76600, None, None], dtype=pl.UInt32))
     assert_frame_equal(df2, df3)
 
     df2 = df.with_columns(tmdb_find(tmdb_type="tv", external_id_type="imdb_id"))
-    df3 = df.with_columns(pl.Series("tmdb_id", [None, 120998, None]))
+    df3 = df.with_columns(pl.Series("tmdb_id", [None, 120998, None], dtype=pl.UInt32))
     assert_frame_equal(df2, df3)
 
     df2 = df.with_columns(tmdb_find(tmdb_type="person", external_id_type="imdb_id"))
-    df3 = df.with_columns(pl.Series("tmdb_id", [None, None, 1674162]))
+    df3 = df.with_columns(pl.Series("tmdb_id", [None, None, 1674162], dtype=pl.UInt32))
     assert_frame_equal(df2, df3)
